Structured for modulation and pytests
