--- passwords.txt
@@ -1,3 +1,0 @@
-Password123!
-Helloworld123!
-Testpass123!
