+++ src/passwords.txt
@@ -1,0 +1,3 @@
+Password123!
+Highfive123!
+Password123!
